tests/wpt/tests/css/filter-effects/parsing/flood-opacity-invalid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>Filter Effects Module Level 1: parsing flood-opacity with invalid values</title>
  <metadata>
    <h:link rel="help" href="https://drafts.fxtf.org/filter-effects/#propdef-flood-opacity"/>
    <h:meta name="assert" content="flood-opacity supports only the grammar 'alpha-value'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_invalid_value("flood-opacity", "1.");
test_invalid_value("flood-opacity", "2 3");

  ]]></script>
</svg>
